refactor: rename styles folder as theme
