+++ src/components/_App/HeadContent.tsx
@@ -1,4 +1,4 @@
-import theme from "../../styles/theme";
+import theme from "../../theme";
 
 const HeadContent = () => (
   <>
