+++ src/pages/_app.tsx
@@ -7,8 +7,8 @@ import client from "../apollo/client";
 import AuthWrapper from "../components/Auth/AuthWrapper";
 import Layout from "../components/_App/Layout";
 import "../i18n/i18n.config";
-import "../styles/globals.css";
-import theme from "../styles/theme";
+import theme from "../theme";
+import "../theme/styles.css";
 import { initAxe } from "../utils/common.utils";
 
 const clientSideEmotionCache = createCache({ key: "css" });
